+++ app/src/js/utils/table-config/execution-status.js
@@ -15,7 +15,47 @@ export const tableColumns = [
   },
   {
     Header: 'Type',
-    accessor: 'type'
+    accessor: 'type',
+    Cell: ({ cell: { value }, row: { original: { eventDetails } } }) => {
+      const [showModal, setShowModal] = useState(false);
+      const { id } = eventDetails || {};
+      function toggleModal(e) {
+        if (e) {
+          e.preventDefault();
+        }
+        setShowModal(!showModal);
+      }
+      if (eventDetails) {
+        let buttonClass;
+        if (eventDetails.type === 'LambdaFunctionFailed') {
+          buttonClass = 'button button--small button--no-left-padding button--failed';
+        } else {
+          buttonClass = 'button button--small button--no-left-padding';
+        }
+        return (
+          <>
+            <button
+              onClick={toggleModal}
+              className={buttonClass}
+            >
+              {eventDetails.type}
+            </button>
+            <DefaultModal
+              showModal={showModal}
+              title={`ID ${id}: ${value}`}
+              onCloseModal={toggleModal}
+              hasConfirmButton={false}
+              cancelButtonClass="button--close"
+              cancelButtonText="Close"
+              className="execution__modal"
+            >
+              <pre>{JSON.stringify(eventDetails, null, 2)}</pre>
+            </DefaultModal>
+          </>
+        );
+      }
+      return 'N/A';
+    },
   },
   {
     Header: 'Step Name',